requirements.txt
pycares==4.4.0
aiodns==3.2.0
langchain-core>=0.1.52
langchain>=0.1.0
langchain-openai>=0.0.8
scrapegraphai>=1.0.0
python-dotenv>=1.0.0
openai>=1.0.0
pandas>=2.0.0
openpyxl>=3.0.0
rich>=13.0.0
streamlit>=1.28.0
plotly>=5.0.0
fastapi>=0.104.0
uvicorn>=0.24.0